IBX-12091: Improved TypeScript translation extractor structure
