+++ src/bundle/Resources/config/services/translation.yaml
@@ -12,6 +12,8 @@ services:
         tags:
             - { name: jms_translation.file_visitor }
 
+    Ibexa\AdminUi\Translation\Extractor\TypeScriptExtractorClient: ~
+
     Ibexa\AdminUi\Translation\Extractor\SortingTranslationExtractor:
         tags:
             - { name: jms_translation.extractor, alias: ez_location_sorting }
