+++ src/lib/Translation/Extractor/TypeScriptFileVisitor.php
@@ -15,41 +15,18 @@ use JMS\TranslationBundle\Model\MessageCatalogue;
 use JMS\TranslationBundle\Translation\Extractor\FileVisitorInterface;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\NullLogger;
-use RuntimeException;
 use SplFileInfo;
-use Symfony\Component\Process\InputStream;
-use Symfony\Component\Process\Process;
 use Twig\Node\Node as TwigNode;
 
 final class TypeScriptFileVisitor implements FileVisitorInterface, LoggerAwareInterface
 {
     use LoggerAwareTrait;
 
-    private const DEFAULT_RESPONSE_TIMEOUT_SECONDS = 30.0;
-
-    private string $parserScriptPath;
-
-    private ?bool $runtimeReady = null;
-
-    private ?Process $serverProcess = null;
-
-    private ?InputStream $serverInput = null;
-
-    private string $serverOutputBuffer = '';
-
     public function __construct(
+        private readonly TypeScriptExtractorClient $extractorClient,
         private readonly string $defaultDomain = 'messages',
-        ?string $parserScriptPath = null,
-        private readonly string $nodeBinary = 'node',
-        private readonly float $responseTimeoutSeconds = self::DEFAULT_RESPONSE_TIMEOUT_SECONDS,
     ) {
         $this->logger = new NullLogger();
-        $this->parserScriptPath = $parserScriptPath ?? __DIR__ . '/typescript_translation_extractor.mjs';
-    }
-
-    public function __destruct()
-    {
-        $this->discardServerProcess();
     }
 
     public function visitFile(SplFileInfo $file, MessageCatalogue $catalogue): void
@@ -58,9 +35,27 @@ final class TypeScriptFileVisitor implements FileVisitorInterface, LoggerAwareIn
             return;
         }
 
-        $this->assertRuntimeIsReady();
+        $messages = $this->extractMessages($file);
 
-        $decoded = $this->requestExtraction((string) $file->getRealPath());
+        if ($messages === null) {
+            return;
+        }
+
+        foreach ($messages as $messageData) {
+            if (!is_array($messageData) || !isset($messageData['id']) || !is_string($messageData['id'])) {
+                continue;
+            }
+
+            $catalogue->add($this->createMessage($file, $messageData));
+        }
+    }
+
+    /**
+     * @return array<mixed>|null
+     */
+    private function extractMessages(SplFileInfo $file): ?array
+    {
+        $decoded = $this->extractorClient->extract((string) $file->getRealPath());
 
         if (isset($decoded['error'])) {
             $this->logger?->error(sprintf(
@@ -69,7 +64,7 @@ final class TypeScriptFileVisitor implements FileVisitorInterface, LoggerAwareIn
                 $decoded['error'],
             ));
 
-            return;
+            return null;
         }
 
         if (!isset($decoded['messages']) || !is_array($decoded['messages'])) {
@@ -78,32 +73,47 @@ final class TypeScriptFileVisitor implements FileVisitorInterface, LoggerAwareIn
                 $file->getRealPath(),
             ));
 
+            return null;
+        }
+
+        $this->logWarnings($decoded['warnings'] ?? []);
+
+        return $decoded['messages'];
+    }
+
+    /**
+     * @param mixed $warnings
+     */
+    private function logWarnings(mixed $warnings): void
+    {
+        if (!is_array($warnings)) {
             return;
         }
 
-        foreach ($decoded['warnings'] ?? [] as $warning) {
+        foreach ($warnings as $warning) {
             if (is_string($warning)) {
                 $this->logger?->error($warning);
             }
         }
+    }
 
-        foreach ($decoded['messages'] as $messageData) {
-            if (!is_array($messageData) || !isset($messageData['id']) || !is_string($messageData['id'])) {
-                continue;
-            }
+    /**
+     * @param array{id: string, domain?: mixed, desc?: mixed} $messageData
+     */
+    private function createMessage(SplFileInfo $file, array $messageData): Message
+    {
+        $message = new Message(
+            $messageData['id'],
+            isset($messageData['domain']) && is_string($messageData['domain']) ? $messageData['domain'] : $this->defaultDomain,
+        );
 
-            $message = new Message(
-                $messageData['id'],
-                isset($messageData['domain']) && is_string($messageData['domain']) ? $messageData['domain'] : $this->defaultDomain,
-            );
-
-            if (isset($messageData['desc']) && is_string($messageData['desc'])) {
-                $message->setDesc($messageData['desc']);
-            }
-
-            $message->addSource(new FileSource((string) $file));
-            $catalogue->add($message);
+        if (isset($messageData['desc']) && is_string($messageData['desc'])) {
+            $message->setDesc($messageData['desc']);
         }
+
+        $message->addSource(new FileSource((string) $file));
+
+        return $message;
     }
 
     /**
@@ -124,144 +134,5 @@ final class TypeScriptFileVisitor implements FileVisitorInterface, LoggerAwareIn
         return ($path !== false)
             && (str_ends_with($path, '.ts') || str_ends_with($path, '.tsx'))
             && !str_ends_with($path, '.d.ts');
-    }
-
-    private function assertRuntimeIsReady(): void
-    {
-        if ($this->runtimeReady === true) {
-            return;
-        }
-
-        if (!is_file($this->parserScriptPath)) {
-            throw new RuntimeException(sprintf(
-                'TypeScript translation extractor script not found: %s.',
-                $this->parserScriptPath,
-            ));
-        }
-
-        $process = new Process([
-            $this->nodeBinary,
-            $this->parserScriptPath,
-            '--check-runtime',
-        ]);
-        $process->run();
-
-        if (!$process->isSuccessful()) {
-            throw new RuntimeException(sprintf(
-                'TypeScript translation extractor runtime is not available. Please ensure `%s` is installed and `@typescript-eslint/typescript-estree` is available for the admin-ui package. %s',
-                $this->nodeBinary,
-                trim($process->getErrorOutput()),
-            ));
-        }
-
-        $this->runtimeReady = true;
-    }
-
-    /**
-     * @return array{messages?: mixed, warnings?: mixed, error?: string}
-     */
-    private function requestExtraction(string $filePath): array
-    {
-        $this->ensureServerStarted();
-
-        $this->getServerInput()->write($filePath . "\n");
-
-        return $this->readServerLine();
-    }
-
-    private function ensureServerStarted(): void
-    {
-        if ($this->serverProcess !== null && $this->serverProcess->isRunning()) {
-            return;
-        }
-
-        $input = new InputStream();
-
-        $process = new Process([
-            $this->nodeBinary,
-            $this->parserScriptPath,
-            '--serve',
-        ]);
-        $process->setInput($input);
-        $process->setTimeout(null);
-        $process->start();
-
-        $this->serverInput = $input;
-        $this->serverProcess = $process;
-        $this->serverOutputBuffer = '';
-    }
-
-    private function getServerInput(): InputStream
-    {
-        if ($this->serverInput === null) {
-            throw new RuntimeException('TypeScript translation extractor input stream has not been started.');
-        }
-
-        return $this->serverInput;
-    }
-
-    private function getServerProcess(): Process
-    {
-        if ($this->serverProcess === null) {
-            throw new RuntimeException('TypeScript translation extractor process has not been started.');
-        }
-
-        return $this->serverProcess;
-    }
-
-    /**
-     * @return array{messages?: mixed, warnings?: mixed, error?: string}
-     */
-    private function readServerLine(): array
-    {
-        $process = $this->getServerProcess();
-        $deadline = microtime(true) + $this->responseTimeoutSeconds;
-
-        while (!str_contains($this->serverOutputBuffer, "\n")) {
-            if (!$process->isRunning()) {
-                $this->discardServerProcess();
-
-                throw new RuntimeException(sprintf(
-                    'TypeScript translation extractor process terminated unexpectedly: %s',
-                    trim($process->getErrorOutput()),
-                ));
-            }
-
-            if (microtime(true) > $deadline) {
-                $this->discardServerProcess();
-
-                throw new RuntimeException(
-                    'Timed out waiting for the TypeScript translation extractor process to respond.',
-                );
-            }
-
-            $this->serverOutputBuffer .= $process->getIncrementalOutput();
-
-            if (!str_contains($this->serverOutputBuffer, "\n")) {
-                usleep(2000);
-            }
-        }
-
-        [$line, $rest] = explode("\n", $this->serverOutputBuffer, 2);
-        $this->serverOutputBuffer = $rest;
-
-        $decoded = json_decode($line, true);
-
-        return is_array($decoded) ? $decoded : ['error' => 'Unable to decode extractor output.'];
-    }
-
-    /**
-     * Kills the current server process and discards any buffered output, so that a
-     * response arriving after a timeout cannot be misread as the answer to a later,
-     * unrelated request once the process is restarted.
-     */
-    private function discardServerProcess(): void
-    {
-        $this->serverInput?->close();
-        $this->serverProcess?->stop(3);
-
-        $this->serverProcess = null;
-        $this->serverInput = null;
-        $this->serverOutputBuffer = '';
     }
 }
