+++ src/lib/Translation/Extractor/typescript_translation_extractor.mjs
@@ -30,11 +30,104 @@ const getStringLiteralValue = (node) => {
 
 const getErrorMessage = (error) => (error instanceof Error ? error.message : String(error));
 
+const TRANSLATOR_ARGUMENTS = {
+    trans: { id: 0, domain: 2 },
+    transChoice: { id: 0, domain: 3 },
+};
+
 const formatWarning = (argumentName, node, filePath) => {
     const line = node?.loc?.start?.line ?? 0;
     const column = node?.loc?.start?.column ?? 0;
 
     return `Could not extract ${argumentName}, expected string literal but got ${node?.type ?? 'nothing'} (in ${filePath} on line ${line} column ${column}).`;
+};
+
+const findClosestLeadingComment = (source, comments, node) => {
+    if (!node?.range) {
+        return null;
+    }
+
+    const [nodeStart] = node.range;
+    let candidate = null;
+
+    for (const comment of comments) {
+        if (!comment.range || comment.range[1] > nodeStart) {
+            continue;
+        }
+
+        const textBetween = source.slice(comment.range[1], nodeStart);
+        if (!/^\s*$/.test(textBetween)) {
+            continue;
+        }
+
+        if (candidate === null || comment.range[1] > candidate.range[1]) {
+            candidate = comment;
+        }
+    }
+
+    return candidate;
+};
+
+const extractDescription = (source, comments, node) => {
+    const comment = findClosestLeadingComment(source, comments, node);
+
+    if (!comment?.value) {
+        return null;
+    }
+
+    const match = comment.value.match(/@Desc\((['"])((?:\\[\s\S]|(?!\1)[\s\S])*)\1\)/);
+
+    return match ? match[2].replace(/\\(['"\\])/g, '$1') : null;
+};
+
+const extractMessage = (node, filePath, source, comments, warnings) => {
+    const methodName = node.callee.property.name;
+    const { id: idArgIndex, domain: domainArgIndex } = TRANSLATOR_ARGUMENTS[methodName];
+    const idNode = node.arguments[idArgIndex];
+    const id = getStringLiteralValue(idNode);
+
+    if (id === null) {
+        warnings.push(formatWarning('id', idNode, filePath));
+        return null;
+    }
+
+    const domainNode = node.arguments[domainArgIndex];
+    const domain = domainNode ? getStringLiteralValue(domainNode) : null;
+
+    if (domainNode && domain === null) {
+        warnings.push(formatWarning('domain', domainNode, filePath));
+    }
+
+    return {
+        id,
+        domain,
+        desc: extractDescription(source, comments, idNode),
+    };
+};
+
+const visitNode = (node, context) => {
+    if (!node || typeof node !== 'object') {
+        return;
+    }
+
+    if (isTranslatorCall(node)) {
+        const message = extractMessage(node, context.filePath, context.source, context.comments, context.warnings);
+
+        if (message !== null) {
+            context.messages.push(message);
+        }
+    }
+
+    for (const value of Object.values(node)) {
+        if (Array.isArray(value)) {
+            value.forEach((child) => visitNode(child, context));
+            continue;
+        }
+
+        if (value && typeof value === 'object') {
+            visitNode(value, context);
+        }
+    }
 };
 
 const extractFromFile = (filePath) => {
@@ -47,92 +140,17 @@ const extractFromFile = (filePath) => {
         filePath,
     });
 
-    const comments = ast.comments ?? [];
-    const messages = [];
-    const warnings = [];
-
-    const findClosestLeadingComment = (node) => {
-        if (!node?.range) {
-            return null;
-        }
-
-        const [nodeStart] = node.range;
-        let candidate = null;
-
-        for (const comment of comments) {
-            if (!comment.range || comment.range[1] > nodeStart) {
-                continue;
-            }
-
-            const textBetween = source.slice(comment.range[1], nodeStart);
-            if (!/^\s*$/.test(textBetween)) {
-                continue;
-            }
-
-            if (candidate === null || comment.range[1] > candidate.range[1]) {
-                candidate = comment;
-            }
-        }
-
-        return candidate;
+    const context = {
+        filePath,
+        source,
+        comments: ast.comments ?? [],
+        messages: [],
+        warnings: [],
     };
 
-    const extractDesc = (node) => {
-        const comment = findClosestLeadingComment(node);
+    visitNode(ast, context);
 
-        if (!comment?.value) {
-            return null;
-        }
-
-        const match = comment.value.match(/@Desc\((['"])((?:\\[\s\S]|(?!\1)[\s\S])*)\1\)/);
-
-        return match ? match[2].replace(/\\(['"\\])/g, '$1') : null;
-    };
-
-    const visit = (node) => {
-        if (!node || typeof node !== 'object') {
-            return;
-        }
-
-        if (isTranslatorCall(node)) {
-            const methodName = node.callee.property.name;
-            const domainArgIndex = methodName === 'trans' ? 2 : 3;
-            const idNode = node.arguments[0];
-            const id = getStringLiteralValue(idNode);
-
-            if (id === null) {
-                warnings.push(formatWarning('id', idNode, filePath));
-            } else {
-                const domainNode = node.arguments[domainArgIndex];
-                const domain = domainNode ? getStringLiteralValue(domainNode) : null;
-
-                if (domainNode && domain === null) {
-                    warnings.push(formatWarning('domain', domainNode, filePath));
-                }
-
-                messages.push({
-                    id,
-                    domain,
-                    desc: extractDesc(idNode),
-                });
-            }
-        }
-
-        for (const value of Object.values(node)) {
-            if (Array.isArray(value)) {
-                value.forEach(visit);
-                continue;
-            }
-
-            if (value && typeof value === 'object') {
-                visit(value);
-            }
-        }
-    };
-
-    visit(ast);
-
-    return { messages, warnings };
+    return { messages: context.messages, warnings: context.warnings };
 };
 
 const [, , mode] = process.argv;
